backend_fastapi/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
openai
google-genai
httpx
whisper
python-multipart
